Added restaurant card
