+++ src/features/restaurants/screens/restaurants.screen.js
@@ -1,0 +1,30 @@
+import React from "react";
+import { StatusBar, StyleSheet, SafeAreaView, Text, View } from "react-native";
+import { Searchbar } from "react-native-paper";
+import { ResturantInfo } from "../components/restaurants-info.component";
+
+export const RestaurantsScreen = () => (
+  <SafeAreaView style={styles.container}>
+    <View style={styles.search}>
+      <Searchbar />
+    </View>
+    <View style={styles.list}>
+      <ResturantInfo />
+    </View>
+  </SafeAreaView>
+);
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    marginTop: StatusBar.currentHeight,
+  },
+  search: {
+    padding: 16,
+  },
+  list: {
+    flex: 1,
+    padding: 16,
+    backgroundColor: "blue",
+  },
+});
